refactor(api): extract notification interfaces to types-notification.ts (#837)
repositories/types.ts sat at 793/800 (the hard file-size cap). Move the
notification_log data-access contract — SEND_LEASE_MS + MAX_NOTIFICATION_ATTEMPTS
consts and the NotificationLogUpsert / NotificationLogFilters /
NotificationLogRepository interfaces — into a sibling types-notification.ts,
mirroring the existing types-payment.ts extraction. types.ts re-exports all
five, so every consumer keeps importing from ../types (zero call-site changes).
types.ts: 793 -> 753 lines. Behavior-preserving.
